+++ packages/graphql-proxy/example/resolvers.js
@@ -1,0 +1,1 @@
+{"Query":{},"Mutation":{}}
