+++ packages/graphql-proxy/package.json
@@ -4,15 +4,17 @@
   "main": "lib/index.js",
   "dependencies": {
     "@workpop/simple-logger": "^1.0.0",
-    "apollo-fetch": "^0.7.0",
+    "apollo-link": "^1.2.1",
+    "apollo-link-http": "^1.5.3",
     "body-parser": "^1.17.2",
+    "command-line-args": "^5.0.2",
     "express": "^4.15.3",
-    "graphql": "^0.11.7",
+    "graphql-binding": "1.2.5",
     "graphql-request": "^1.2.0",
-    "graphql-server-express": "^1.2.0",
-    "graphql-tools": "^2.11.0",
+    "graphql-server-express": "^1.3.4",
+    "graphql-tools": "^2.23.1",
     "lodash": "^4.17.4",
-    "node-fetch": "^1.7.1"
+    "node-fetch": "^2.1.2"
   },
   "devDependencies": {
     "@workpop/eslint-config-workpop": "^0.0.9",
@@ -39,5 +41,8 @@
     "lint:fix": "eslint . --fix",
     "testonly": "mocha ./__tests__/*.test.js --compilers js:babel-register",
     "test:watch": "npm run testonly -- --watch --watch-extensions js"
+  },
+  "peerDependencies": {
+    "graphql": "0.13.2"
   }
 }
