+++ packages/graphql-proxy/src/ServiceTransport.js
@@ -1,0 +1,12 @@
+import { HttpLink } from 'apollo-link-http';
+import fetch from 'node-fetch';
+
+export default class ServiceLink extends HttpLink {
+  constructor({ headers, uri }) {
+    super({
+      uri,
+      headers,
+      fetch,
+    });
+  }
+}
